requirements.txt
python_documentcloud==4.0.1
azure-ai-formrecognizer
